+++ backend/timetable-back/src/main/java/ru/nsu/timetable/services/TeacherService.java
@@ -3,33 +3,51 @@ package ru.nsu.timetable.services;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.nsu.timetable.exceptions.ResourceNotFoundException;
+import ru.nsu.timetable.models.dto.TeacherDTO;
 import ru.nsu.timetable.models.entities.Teacher;
+import ru.nsu.timetable.models.mappers.TeacherMapper;
 import ru.nsu.timetable.repositories.TeacherRepository;
 
+@RequiredArgsConstructor
 @Service
 public class TeacherService {
     private final TeacherRepository teacherRepository;
+    private final TeacherMapper teacherMapper;
 
-    @Autowired
-    public TeacherService(TeacherRepository teacherRepository) {
-        this.teacherRepository = teacherRepository;
+    public List<TeacherDTO> getAllTeachers() {
+        return teacherRepository
+                .findAll()
+                .stream()
+                .map(teacherMapper::toTeacherDTO)
+                .toList();
     }
 
-    public List<Teacher> getAllTeachers() {
-        return teacherRepository.findAll();
+    public TeacherDTO getTeacherById(Long id) {
+        return teacherMapper.toTeacherDTO(getTeacher(id));
     }
 
-    public Optional<Teacher> getTeacherById(Long id) {
-        return teacherRepository.findById(id);
-    }
-
-    public Teacher saveTeacher(Teacher teacher) {
-        return teacherRepository.save(teacher);
+    public TeacherDTO saveTeacher(TeacherDTO teacherDTO) {
+        Teacher course = teacherMapper.toTeacher(teacherDTO);
+        return teacherMapper.toTeacherDTO(teacherRepository.save(course));
     }
 
     public void deleteTeacher(Long id) {
-        teacherRepository.deleteById(id);
+        if (teacherRepository.existsById(id)) {
+            teacherRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Course with id " + id + " not found");
+        }
+    }
+
+    private Teacher getTeacher(Long id) {
+        Optional<Teacher> teacher = teacherRepository.findById(id);
+        if (teacher.isEmpty()) {
+            throw new ResourceNotFoundException("Course with id " + id + " not found");
+        } else {
+            return teacher.get();
+        }
     }
 }
